added plugin installation
improved build.sh build helper.

+++ build.sh
@@ -1,6 +1,10 @@
-rm -rf build
+#!/bin/bash
+if [[ $1 = "--clean" ]]
+then
+    rm -rf build
+fi
 rm -rf BUILDROOT
-cmake -S . -B build
-cmake --build build
+DESTDIR=BUILDROOT cmake -S . -B build 
+DESTDIR=BUILDROOT cmake --build build
 mkdir -p BUILDROOT
-DESTDIR=BUILDROOT cmake --install build --prefix=/usr
+DESTDIR=BUILDROOT cmake --install build --prefix=/usr
